Moved the bl_info dictionary from add_objects.py to __init__.py.
This is the first step to making this add-on a multi-file add-on.
Alternatively, `add_objects.py` could simply be renamed to __init__.py,
but this example is meant to step through the process more methodically.

bl_info is what Blender uses to populate the Add-ons list in
Preferences. However, if you try to enable the add-on from this commit,
it will error out.